+++ lib/src/exports/size_config.dart
@@ -35,7 +35,7 @@ class SizeConfig {
 ///Adding extension over the double data type for easy access.
 ///
 extension SizeConfigForDouble on num {
-  Size _mq(BuildContext ctx) => MediaQuery.of(ctx).size;
+  Size _mq(BuildContext context) => MediaQuery.of(context).size;
 
   ///The height parameter `h` is used to access the percent height of the current device.
   /// ```
@@ -49,7 +49,7 @@ extension SizeConfigForDouble on num {
   ///everytime when context changes while the latter returns a fixed height
   ///even if screen height changes.
   ///ch is helpful in flutter web as windows are adjustable.
-  double ch(BuildContext ctx) => this * (_mq(ctx).height / 100);
+  double ch(BuildContext context) => this * (_mq(context).height / 100);
 
   ///The width parameter `w` is used to access the percent width of the current device.
   /// ```
@@ -63,5 +63,16 @@ extension SizeConfigForDouble on num {
   ///everytime when context changes while the latter returns a fixed width
   ///even if screen width changes.
   ///cw is helpful in flutter web as windows are adjustable.
-  double cw(BuildContext ctx) => this * (_mq(ctx).width / 100);
+  double cw(BuildContext context) => this * (_mq(context).width / 100);
+
+  /// The `Scale independent pixels (sp)` parameter is a getter method that calculates the
+  /// scaled pixel value based on the current device's width.
+  double get sp => this * (SizeConfig._width / 3) / 100;
+
+  ///csp stands for current Scale independent pixels
+  ///The only difference `num.csp()` and `num.sp` is that [num.csp()] rebuilds
+  ///everytime when context changes while the latter returns a fixed width
+  ///even if screen width changes.
+  ///csp is helpful in flutter web as windows are adjustable.
+  double csp(BuildContext context) => this * (_mq(context).width / 3) / 100;
 }